+++ artemis-weaver/src/test/java/com/artemis/component/PooledNotForced.java
@@ -1,6 +1,7 @@
 package com.artemis.component;
 
 import com.artemis.annotations.PooledWeaver;
+import com.artemis.utils.IntBag;
 
 
 @PooledWeaver()